Fix rehome() for dataclasses and move to cij.util
