+++ modules/cij/reporter.py
@@ -9,6 +9,7 @@ import os
 import jinja2
 import cij.runner
 import cij
+from cij.util import rehome
 
 
 def extract_hook_names(ent):
@@ -243,29 +244,6 @@ def dset_to_html(dset, tmpl_fpath):
     tmpl = env.get_template(tmpl_fname)
 
     return tmpl.render(dset=dset)
-
-
-def rehome(old, new, struct):
-    """
-    Replace all absolute paths to "re-home" it
-    """
-
-    if old == new:
-        return
-
-    if isinstance(struct, list):
-        for item in struct:
-            rehome(old, new, item)
-    elif isinstance(struct, dict):
-        for key, val in struct.iteritems():
-            if isinstance(val, (dict, list)):
-                rehome(old, new, val)
-            elif "conf" in key:
-                continue
-            elif "orig" in key:
-                continue
-            elif "root" in key or "path" in key:
-                struct[key] = struct[key].replace(old, new)
 
 
 def main(args):
